feat: :sparkles: Session Management Checks/prepared for Spring Sec 6 defualt behaviors
Our app didn't break when we did both: 1. Require Explicit Saving of `SecurityContextRepository` 2. Require Explicit Invocation of `SessionAuthenticationStrategy`  these are defaults behaviors in Spring Security 6, the code added here is to be removed after migrating to 6 as this is the default behavior. app running and test are passed proof we are good to go with session management in Spring Security 6.